+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.21.0] - 2026-06-15
+
 ### Added
 
 - **`hpm registry add --if-not-exists`.** Adding a registry whose name